requirements.txt
# Core dependencies
python>=3.10

# Data processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0

# Elasticsearch
elasticsearch>=8.9.0
elastic-transport>=8.4.0

# Vector search & embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4

# Sparse retrieval
splade @ git+https://github.com/naver/splade.git
rank-bm25>=0.2.2

# Reranking
pygaggle>=0.0.3.1

# LLM & prompt engineering
openai>=1.0.0
anthropic>=0.7.0
langchain>=0.1.0
tiktoken>=0.5.0

# Evaluation
evaluate>=0.4.0
rouge-score>=0.1.2
bert-score>=0.3.13
ragas>=0.1.0

# Experiment tracking
mlflow>=2.8.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
rich>=13.0.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0
plotly>=5.14.0
seaborn>=0.12.0
matplotlib>=3.7.0
